requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
PyPDF2==3.0.1
pdfplumber==0.9.0
pytesseract==0.3.10
Pillow==10.0.1
openpyxl==3.1.2
reportlab==4.0.4
python-dateutil==2.8.2
